Split do_poll_tasks and stuff into its own source file.
