+++ brain/CMakeLists.txt
@@ -21,12 +21,12 @@ include_directories(${GAZEBO_INCLUDE_DIRS})
 link_directories(${GAZEBO_LIBRARY_DIRS})
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${GAZEBO_CXX_FLAGS}")
-add_executable(kickerbot main_kickerbot.cc robot.cc common.cpp tasklib.cpp pid.cpp)
+add_executable(kickerbot main_kickerbot.cc robot.cc common.cpp tasklib.cpp taskrt.cpp pid.cpp)
 target_link_libraries(kickerbot ${GTK3_LIBRARIES})
 target_link_libraries(kickerbot ${GAZEBO_LIBRARIES})
 target_link_libraries(kickerbot ${OpenCV_LIBS})
 
-add_executable(goaliebot main_goaliebot.cc robot.cc common.cpp tasklib.cpp pid.cpp)
+add_executable(goaliebot main_goaliebot.cc robot.cc common.cpp tasklib.cpp taskrt.cpp pid.cpp)
 target_link_libraries(goaliebot ${GTK3_LIBRARIES})
 target_link_libraries(goaliebot ${GAZEBO_LIBRARIES})
 target_link_libraries(goaliebot ${OpenCV_LIBS})
