+++ brain/main_kickerbot.cc
@@ -9,6 +9,7 @@
 #include "robot.hh"
 #include "common.hxx"
 #include "tasklib.hxx"
+#include "taskrt.hxx"
 
 using namespace cv;
 using namespace std;
