+++ brain/tasklib.hxx
@@ -76,8 +76,4 @@ public:
 //class DebugPrintPosTask;
 //class DebugPrintLidarTask;
 
-void print_stack_trace(aistate* state);
-void do_abort(Robot* robo, aistate* ai);
-void do_poll_tasks(Robot* robo, aistate* ai);
-
 #endif
